atualiza modelos ConsumoDia e Consumo para volume ser Integer positivo

+++ consumo/models.py
@@ -7,12 +7,12 @@ class ConsumoDia(models.Model):
     perfil = models.ForeignKey(Perfil, on_delete=models.CASCADE)
     data = models.DateField()
     is_meta_atingida = models.BooleanField(default=False)
-    consumo = models.DecimalField(max_digits=6, decimal_places=2)
+    consumo = models.PositiveIntegerField()
 
     def __str__(self) -> str:
         return self.data.strftime("%Y-%m-%d")
 
 class Consumo(models.Model):
     consumoDia = models.ForeignKey(ConsumoDia, on_delete=models.CASCADE, related_name="consumos", null=True)
-    volume = models.DecimalField(max_digits=6, decimal_places=2)
+    volume = models.PositiveIntegerField()
     data = models.DateTimeField(auto_now_add=True)
